Added unit test to .NET Core test suite. Issue #538
