+++ Tests/Desktop/Network/DicomServiceTest.cs
@@ -1,17 +1,12 @@
 ﻿// Copyright (c) 2012-2017 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
+using System.Threading.Tasks;
+
+using Xunit;
+
 namespace Dicom.Network
 {
-    using System;
-    using System.IO;
-    using System.Text;
-    using System.Threading.Tasks;
-
-    using Dicom.Log;
-
-    using Xunit;
-
     [Collection("Network"), Trait("Category", "Network")]
     public class DicomServiceTest
     {
@@ -73,81 +68,4 @@ namespace Dicom.Network
 
         #endregion
     }
-
-    #region Support classes
-
-    internal class SimpleCStoreProvider : DicomService, IDicomServiceProvider, IDicomCStoreProvider
-    {
-        private static readonly DicomTransferSyntax[] AcceptedTransferSyntaxes =
-            {
-                    DicomTransferSyntax.ExplicitVRLittleEndian,
-                    DicomTransferSyntax.ExplicitVRBigEndian,
-                    DicomTransferSyntax.ImplicitVRLittleEndian
-                };
-
-        private static readonly DicomTransferSyntax[] AcceptedImageTransferSyntaxes =
-            {
-                    // Lossless
-                    DicomTransferSyntax.JPEGLSLossless,
-                    DicomTransferSyntax.JPEG2000Lossless,
-                    DicomTransferSyntax.JPEGProcess14SV1,
-                    DicomTransferSyntax.JPEGProcess14,
-                    DicomTransferSyntax.RLELossless,
-
-                    // Lossy
-                    DicomTransferSyntax.JPEGLSNearLossless,
-                    DicomTransferSyntax.JPEG2000Lossy,
-                    DicomTransferSyntax.JPEGProcess1,
-                    DicomTransferSyntax.JPEGProcess2_4,
-
-                    // Uncompressed
-                    DicomTransferSyntax.ExplicitVRLittleEndian,
-                    DicomTransferSyntax.ExplicitVRBigEndian,
-                    DicomTransferSyntax.ImplicitVRLittleEndian
-                };
-
-        public SimpleCStoreProvider(INetworkStream stream, Encoding fallbackEncoding, Logger log)
-            : base(stream, fallbackEncoding, log)
-        {
-        }
-
-        public Task OnReceiveAssociationRequestAsync(DicomAssociation association)
-        {
-            foreach (var pc in association.PresentationContexts)
-            {
-                if (pc.AbstractSyntax == DicomUID.Verification) pc.AcceptTransferSyntaxes(AcceptedTransferSyntaxes);
-                else if (pc.AbstractSyntax.StorageCategory != DicomStorageCategory.None) pc.AcceptTransferSyntaxes(AcceptedImageTransferSyntaxes);
-            }
-
-            return SendAssociationAcceptAsync(association);
-        }
-
-        public Task OnReceiveAssociationReleaseRequestAsync()
-        {
-            return SendAssociationReleaseResponseAsync();
-        }
-
-        public void OnReceiveAbort(DicomAbortSource source, DicomAbortReason reason)
-        {
-        }
-
-        public void OnConnectionClosed(Exception exception)
-        {
-        }
-
-        public DicomCStoreResponse OnCStoreRequest(DicomCStoreRequest request)
-        {
-            var tempName = Path.GetTempFileName();
-            Logger.Info(tempName);
-            request.File.Save(tempName);
-
-            return new DicomCStoreResponse(request, DicomStatus.Success);
-        }
-
-        public void OnCStoreRequestException(string tempFileName, Exception e)
-        {
-        }
-    }
-
-    #endregion
 }
